+++ pkg/tool/goqueryd/integration_test/service/search_test.go
@@ -15,7 +15,7 @@ func TestSearchKeyword(t *testing.T) {
 	o := search_option.New("hybrid search pipeline", 10)
 	o.Mode = "keyword"
 	outcome := s.Service.Search(o)
-	assert.Greater(t, 0, float64(len(outcome.Results)))
+	assert.Greater(t, 0, len(outcome.Results))
 	assert.String(t, "Search Pipeline", outcome.Results[0].Title)
 }
 
@@ -25,7 +25,7 @@ func TestSearchHybrid(t *testing.T) {
 	_, e := s.Service.Embed()
 	assert.FatalOnError(t, e)
 	outcome := s.Service.Search(search_option.New("semantic meaning", 10))
-	assert.Greater(t, 0, float64(len(outcome.Results)))
+	assert.Greater(t, 0, len(outcome.Results))
 }
 
 func TestSearchCollectionFilter(t *testing.T) {
@@ -35,7 +35,7 @@ func TestSearchCollectionFilter(t *testing.T) {
 	o.Collection = "test"
 	o.Mode = "keyword"
 	outcome := s.Service.Search(o)
-	assert.Greater(t, 0, float64(len(outcome.Results)))
+	assert.Greater(t, 0, len(outcome.Results))
 	assert.String(t, "test", outcome.Results[0].Collection)
 }
 
@@ -44,5 +44,5 @@ func TestSearchDegradedWithoutEmbeddings(t *testing.T) {
 	s.IndexFixtures()
 	outcome := s.Service.Search(search_option.New("chunking strategy", 10))
 	assert.True(t, outcome.Degraded)
-	assert.Greater(t, 0, float64(len(outcome.Results)))
+	assert.Greater(t, 0, len(outcome.Results))
 }
